+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Parses correctly image mime types (`image/png`, ...) instead of returning `none` when using `Mime.parse('image/png')`.
+
 ## [0.4.0] - 2026-05-15
 
 ### Breaking